requirements.txt
torch
torchvision
segmentation-models-pytorch
numpy
opencv-python
Pillow
albumentations
tqdm
pyyaml
matplotlib
scikit-learn
